Complete docstrings (#236)
* Docstrings in renten_anspr_m.

* Docstrings in demographic_vars.

* Docstrings in taxes except zu_verst_eink.

* Some tax docstrings.

* Done with taxes docstrings.

* Deleted beitr_bemess_grenze in arbeitsl_geld.py. Added docstrings there.

* Refactored soz_vers/.

* Docstrings for elterngeld and unterhaltsvorschuss. Minor corrections in code.

* Created docstrings for kindezruschlag.

* Created docstrings for arbeitsl_geld_2.

* Created docstrings for benefit checks.

* Wrote docstrings for housing benefit.

* Fixed last documentation issues.

* Ran the how-to guide.

* Inserted forgottten docstrings. Startet on output variables.

* Included output variables.

* Inserted first interface documentation.

* Done with interface.
