+++ docs/gettsim_objects/index.rst
@@ -6,7 +6,9 @@ This section documents the code inside GETTSIM.
 .. toctree::
    :maxdepth: 1
 
-   gettsim_objects/functions
-   gettsim_objects/variables
-   gettsim_objects/types
-   gettsim_objects/params
+   interface
+   functions
+   variables
+   variables_out
+   types
+   params
