+++ docs/gettsim_objects/interface.rst
@@ -1,0 +1,11 @@
+The interface of GETTSIM
+=========================
+
+This section provides the documentation of the two interface functions. If you want
+to have more information on how they work and how you can use them please see
+:ref:`how_to` and :ref:`tutorials`.
+
+.. currentmodule:: gettsim.interface
+.. autofunction:: compute_taxes_and_transfers
+.. currentmodule:: gettsim.policy_environment
+.. autofunction:: set_up_policy_environment
