+++ docs/gettsim_objects/variables_out.rst
@@ -1,0 +1,44 @@
+Typical output variables
+========================
+
+The table below gives an overview of typical variables selected by users to simulate
+with GETTSIM. You can simulate them by defining them as target in
+:func:`compute_taxes_and_transfers`. You can find their individual calculation in the
+documentation of all :ref:`functions`.
+
+
++-------------------------+-----------------------------------------------------------+
+| Variables               | Description                                               |
++=========================+===========================================================+
+| rentenv_beitr_m         | Monthly amount employee old-age pensions contributions    |
++-------------------------+-----------------------------------------------------------+
+| arbeitsl_v_beitr_m      | Monthly amount employee unempl. insurance contributions   |
++-------------------------+-----------------------------------------------------------+
+| ges_krankenv_beitr_m    | Monthly amount employee health insurance contributions    |
++-------------------------+-----------------------------------------------------------+
+| pflegev_beitr_m         | Monthly amount of long term care insurance                |
++-------------------------+-----------------------------------------------------------+
+| arbeitsl_geld_m         | Monthly amount of unemployment assistance                 |
++-------------------------+-----------------------------------------------------------+
+| rente_anspr_m           | Monthly amount of old-age pension claim                   |
++-------------------------+-----------------------------------------------------------+
+| entgeltpunkte_update    | Updated earning points for pension claim                  |
++-------------------------+-----------------------------------------------------------+
+| abgelt_st_tu            | Capital income tax per tax unit                           |
++-------------------------+-----------------------------------------------------------+
+| soli_st_tu              | Solidarity surcharge per tax unit                         |
++-------------------------+-----------------------------------------------------------+
+| kindergeld_m            | Monthly child benefit                                     |
++-------------------------+-----------------------------------------------------------+
+| eink_st_tu              | Income Tax per tax unit                                   |
++-------------------------+-----------------------------------------------------------+
+| unterhaltsvors_m        | Alimony advance payment                                   |
++-------------------------+-----------------------------------------------------------+
+| arbeitsl_geld_2_m_hh    | Monthly subsistence payment per household                 |
++-------------------------+-----------------------------------------------------------+
+| kinderzuschlag_m        | Monthly additional child benefit, household sum           |
++-------------------------+-----------------------------------------------------------+
+| elterngeld_m            | Monthly parental leave benefit                            |
++-------------------------+-----------------------------------------------------------+
+| wohngeld_m_hh           | Monthly housing benefit per household                     |
++-------------------------+-----------------------------------------------------------+
