+++ docs/how_to_guides/index.rst
@@ -1,3 +1,5 @@
+.. _how_to:
+
 How-To Guides
 =============
