+++ docs/tutorials/index.rst
@@ -1,3 +1,5 @@
+.. _tutorials:
+
 Tutorials
 =========
