+++ gettsim/benefits/benefit_checks/benefit_checks.py
@@ -3,17 +3,17 @@ from gettsim.typing import FloatSeries
 
 
 def arbeitsl_geld_2_m_minus_eink_hh(
-    _regelbedarf_m_vermögens_check_hh: FloatSeries,
+    regelbedarf_m_vermögens_check_hh: FloatSeries,
     kindergeld_m_hh: FloatSeries,
     unterhaltsvors_m_hh: FloatSeries,
     arbeitsl_geld_2_eink_hh: FloatSeries,
 ) -> FloatSeries:
-    """
+    """Calcualte remaining basic subsistence after recieving other benefits.
 
     Parameters
     ----------
-    _regelbedarf_m_vermögens_check_hh
-        See :func:`_regelbedarf_m_vermögens_check_hh`.
+    regelbedarf_m_vermögens_check_hh
+        See :func:`regelbedarf_m_vermögens_check_hh`.
     kindergeld_m_hh
         See :func:`kindergeld_m_hh`.
     unterhaltsvors_m_hh
@@ -26,7 +26,7 @@ def arbeitsl_geld_2_m_minus_eink_hh(
 
     """
     out = (
-        _regelbedarf_m_vermögens_check_hh
+        regelbedarf_m_vermögens_check_hh
         - arbeitsl_geld_2_eink_hh
         - unterhaltsvors_m_hh
         - kindergeld_m_hh
@@ -38,7 +38,7 @@ def wohngeld_vorrang_hh(
     wohngeld_vermögens_check_hh: FloatSeries,
     arbeitsl_geld_2_m_minus_eink_hh: FloatSeries,
 ) -> BoolSeries:
-    """
+    """Check if housing benefit has priority.
 
     Parameters
     ----------
@@ -58,7 +58,7 @@ def kinderzuschlag_vorrang_hh(
     kinderzuschlag_vermögens_check_hh: FloatSeries,
     arbeitsl_geld_2_m_minus_eink_hh: FloatSeries,
 ) -> BoolSeries:
-    """
+    """Check if child benefit has priority.
 
     Parameters
     ----------
@@ -79,7 +79,7 @@ def wohngeld_kinderzuschlag_vorrang_hh(
     kinderzuschlag_vermögens_check_hh: FloatSeries,
     arbeitsl_geld_2_m_minus_eink_hh: FloatSeries,
 ) -> BoolSeries:
-    """
+    """Check if housing and child benefit have priority.
 
     Parameters
     ----------
