+++ gettsim/soz_vers/arbeitsl_v.py
@@ -3,7 +3,6 @@ import numpy as np
 
 from gettsim.typing import BoolSeries
 from gettsim.typing import FloatSeries
-from gettsim.typing import IntSeries
 
 
 def sozialv_beitr_m(
@@ -31,60 +30,6 @@ def sozialv_beitr_m(
 
     """
     return pflegev_beitr_m + ges_krankenv_beitr_m + rentenv_beitr_m + arbeitsl_v_beitr_m
-
-
-def rentenv_beitr_m(
-    geringfügig_beschäftigt: BoolSeries,
-    rentenv_beitr_regular_job: FloatSeries,
-    an_beitr_rentenv_midi_job: FloatSeries,
-) -> FloatSeries:
-    """Contribution for each individual to the pension insurance.
-
-    Parameters
-    ----------
-    geringfügig_beschäftigt
-        See :func:`geringfügig_beschäftigt`.
-
-    rentenv_beitr_regular_job
-        See :func:`rentenv_beitr_regular_job`.
-
-    an_beitr_rentenv_midi_job
-        See :func:`an_beitr_rentenv_midi_job`.
-
-
-    Returns
-    -------
-
-    """
-
-    out = geringfügig_beschäftigt.astype(float) * np.nan
-
-    # Set to 0 for minijobs
-    out.loc[geringfügig_beschäftigt] = 0
-
-    # Assign calculated contributions
-    out.loc[an_beitr_rentenv_midi_job.index] = an_beitr_rentenv_midi_job
-    out.loc[rentenv_beitr_regular_job.index] = rentenv_beitr_regular_job
-
-    return out
-
-
-def rentenv_beitr_m_tu(rentenv_beitr_m: FloatSeries, tu_id: IntSeries) -> FloatSeries:
-    """Calculate the contribution of each tax unit to the pension insurance.
-
-    Parameters
-    ----------
-    rentenv_beitr_m
-        See :func:`rentenv_beitr_m`.
-
-    tu_id
-        See basic input variable :ref:`tu_id <tu_id>`.
-
-    Returns
-    -------
-
-    """
-    return rentenv_beitr_m.groupby(tu_id).sum()
 
 
 def arbeitsl_v_beitr_m(
@@ -144,29 +89,6 @@ def arbeitsl_v_regular_job(
     )
 
 
-def rentenv_beitr_regular_job(
-    ges_beitr_arbeitsl_v_midi_jobreturn: FloatSeries, soz_vers_beitr_params: dict
-) -> FloatSeries:
-    """Calculates pension insurance contributions for regular jobs.
-
-    Parameters
-    ----------
-    ges_beitr_arbeitsl_v_midi_jobreturn
-        See :func:`ges_beitr_arbeitsl_v_midi_jobreturn`.
-
-    soz_vers_beitr_params
-        See params documentation :ref:`soz_vers_beitr_params <soz_vers_beitr_params>`.
-
-    Returns
-    -------
-
-    """
-    return (
-        ges_beitr_arbeitsl_v_midi_jobreturn
-        * soz_vers_beitr_params["soz_vers_beitr"]["rentenv"]
-    )
-
-
 def ges_beitr_arbeitsl_v_midi_jobreturn(
     bruttolohn_m: FloatSeries,
     rentenv_beitr_bemess_grenze: FloatSeries,
@@ -195,29 +117,6 @@ def ges_beitr_arbeitsl_v_midi_jobreturn(
     return bruttolohn_m_regulär_beschäftigt.clip(upper=bemess_grenze)
 
 
-def rentenv_beitr_bemess_grenze(
-    wohnort_ost: BoolSeries, soz_vers_beitr_params: dict
-) -> FloatSeries:
-    """
-
-    Parameters
-    ----------
-    wohnort_ost
-    soz_vers_beitr_params
-
-    Returns
-    -------
-
-    """
-    out = wohnort_ost.replace(
-        {
-            True: soz_vers_beitr_params["beitr_bemess_grenze"]["rentenv"]["ost"],
-            False: soz_vers_beitr_params["beitr_bemess_grenze"]["rentenv"]["west"],
-        }
-    )
-    return out.astype(float)
-
-
 def ges_beitr_arbeitsl_v_midi_job(
     midi_job_bemessungsentgelt: FloatSeries, soz_vers_beitr_params: dict
 ) -> FloatSeries:
@@ -240,57 +139,6 @@ def ges_beitr_arbeitsl_v_midi_job(
         * 2
         * soz_vers_beitr_params["soz_vers_beitr"]["arbeitsl_v"]
     )
-
-
-def ges_beitr_rentenv_midi_job(
-    midi_job_bemessungsentgelt: FloatSeries, soz_vers_beitr_params: dict
-) -> FloatSeries:
-    """Calculating the sum of employee and employer pension insurance contribution.
-
-    Parameters
-    ----------
-    midi_job_bemessungsentgelt
-        See :func:`midi_job_bemessungsentgelt`.
-
-    soz_vers_beitr_params
-        See params documentation :ref:`soz_vers_beitr_params <soz_vers_beitr_params>`.
-
-    Returns
-    -------
-
-    """
-    return (
-        midi_job_bemessungsentgelt
-        * 2
-        * soz_vers_beitr_params["soz_vers_beitr"]["rentenv"]
-    )
-
-
-def ag_beitr_rentenv_midi_job(
-    bruttolohn_m: FloatSeries, in_gleitzone: BoolSeries, soz_vers_beitr_params: dict
-) -> FloatSeries:
-    """Calculating the employer pension insurance contribution.
-
-    Parameters
-    ----------
-    bruttolohn_m
-        See basic input variable :ref:`bruttolohn_m <bruttolohn_m>`.
-
-    in_gleitzone
-        See :func:`in_gleitzone`.
-
-    soz_vers_beitr_params
-        See params documentation :ref:`soz_vers_beitr_params <soz_vers_beitr_params>`.
-
-    Returns
-    -------
-
-    """
-    bruttolohn_m__in_gleitzone = bruttolohn_m.loc[in_gleitzone]
-    out = (
-        bruttolohn_m__in_gleitzone * soz_vers_beitr_params["soz_vers_beitr"]["rentenv"]
-    )
-    return out
 
 
 def ag_beitr_arbeitsl_v_midi_job(
@@ -318,26 +166,6 @@ def ag_beitr_arbeitsl_v_midi_job(
         bruttolohn_m_in_gleitzone
         * soz_vers_beitr_params["soz_vers_beitr"]["arbeitsl_v"]
     )
-
-
-def an_beitr_rentenv_midi_job(
-    ges_beitr_rentenv_midi_job: FloatSeries, ag_beitr_rentenv_midi_job: FloatSeries
-) -> FloatSeries:
-    """Calculating the employer unemployment insurance contribution.
-
-    Parameters
-    ----------
-    ges_beitr_rentenv_midi_job
-        See :func:`ges_beitr_rentenv_midi_job`.
-
-    ag_beitr_rentenv_midi_job
-        See :func:`ag_beitr_rentenv_midi_job`.
-
-    Returns
-    -------
-
-    """
-    return ges_beitr_rentenv_midi_job - ag_beitr_rentenv_midi_job
 
 
 def an_beitr_arbeitsl_v_midi_job(
