+++ gettsim/tests/test_data/test_dfs_alg2.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","vorstand_tu","kind","alter","kaltmiete_m_hh","heizkosten_m_hh","wohnfläche_hh","bewohnt_eigentum_hh","alleinerziehend","bruttolohn_m","ges_rente_m","kapital_eink_m","arbeitsl_geld_m","sonstig_eink_m","eink_selbst_m","vermiet_eink_m","eink_st_tu","soli_st_tu","sozialv_beitr_m","kindergeld_m_hh","unterhaltsvors_m","elterngeld_m","vermögen_hh","jahr","sum_basis_arbeitsl_geld_2_eink","arbeitsl_geld_2_eink_hh","arbeitsl_geld_2_eink","alleinerziehenden_mehrbedarf_hh","regelsatz_m_hh","kost_unterk_m_hh","regelbedarf_m_hh","arbeitsl_geld_2_eink_old","eink_anr_frei","vermfreibetr","maxvermfb","cost_per_qm","wohnfl_just","hhsize","unterhaltsvors_m_hh","_arbeitsl_geld_2_brutto_eink_hh","nettoquote",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
+"hh_id","tu_id","p_id","vorstand_tu","kind","alter","kaltmiete_m_hh","heizkosten_m_hh","wohnfläche_hh","bewohnt_eigentum_hh","alleinerziehend","bruttolohn_m","ges_rente_m","kapital_eink_m","arbeitsl_geld_m","sonstig_eink_m","eink_selbst_m","vermiet_eink_m","eink_st_tu","soli_st_tu","sozialv_beitr_m","kindergeld_m_hh","unterhaltsvors_m","elterngeld_m","vermögen_hh","jahr","sum_basis_arbeitsl_geld_2_eink","arbeitsl_geld_2_eink_hh","arbeitsl_geld_2_eink","alleinerziehenden_mehrbedarf_hh","regelsatz_m_hh","kost_unterk_m_hh","regelbedarf_m_hh","arbeitsl_geld_2_eink_old","eink_anr_frei","vermfreibetr","maxvermfb","cost_per_qm","wohnfl_just","hhsize","unterhaltsvors_m_hh","arbeitsl_geld_2_brutto_eink_hh","nettoquote",,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,,
 2,2,4,"TRUE","FALSE",40,400,80,70,"TRUE","FALSE",800,0,0,0,100,0,0,0,0,160,0,0,0,0,2013,500,500,500,0,690,480,1170,740,240,,,6.85714285714286,70,2,0,900,,1,1,1,"TRUE","FALSE",30,700,100,75,"FALSE","FALSE",1000,0,0,0,0,0,0,1200,240,200,190,0,0,0,2016,630,440,400,0,965,750,1715,680,280,,,10,75,3,0,1200,
 2,2,5,"FALSE","FALSE",40,400,80,70,"TRUE","FALSE",0,0,0,0,0,0,0,0,0,0,0,0,0,0,2013,500,500,0,0,690,480,1170,0,0,,,6.85714285714286,70,2,0,900,,1,12,2,"FALSE","FALSE",30,700,100,75,"FALSE","FALSE",0,0,0,0,0,0,0,0,0,0,190,0,0,0,2016,630,440,0,0,965,750,1715,0,0,,,10,75,3,0,1200,
 3,3,6,"TRUE","FALSE",28,400,80,58,"FALSE","TRUE",1000,0,0,0,250,0,0,1200,66,200,164,0,0,0,2009,848.5,684.5,684.5,0.36,687.96,480,1167.96,944.5,260,,,8.27586206896552,58,2,0,1250,,1,1,3,"FALSE","TRUE",2,700,100,75,"FALSE","FALSE",200,0,0,0,0,0,0,0,0,40,190,0,0,0,2016,630,440,40,0,965,750,1715,160,120,,,10,75,3,0,1200,
