+++ gettsim/tests/test_data/test_dfs_kiz.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","head","hhtyp","hh_korr","hhsize","kind","kindergeld_anspruch","alter","arbeitsstunden_w","bruttolohn_m","in_ausbildung","kaltmiete_m_hh","heizkosten_m_hh","alleinerziehend","alleinerziehenden_mehrbedarf_hh","anz_erw_tu","anz_kinder_tu","_arbeitsl_geld_2_brutto_eink_hh","arbeitsl_geld_2_eink_hh","kindergeld_m_hh","unterhaltsvors_m","jahr","kinderzuschlag_m_vorläufig",,,,"kinderzuschlag_eink_regel","kinderzuschlag_kosten_unterk_m",,"kinderzuschlag_eink_regel","kinderzuschlag_eink_max","kinderzuschlag_eink_spanne","kinderzuschlag_eink_anrechn","kinderzuschlag_temp"
+"hh_id","tu_id","p_id","head","hhtyp","hh_korr","hhsize","kind","kindergeld_anspruch","alter","arbeitsstunden_w","bruttolohn_m","in_ausbildung","kaltmiete_m_hh","heizkosten_m_hh","alleinerziehend","alleinerziehenden_mehrbedarf_hh","anz_erw_tu","anz_kinder_tu","arbeitsl_geld_2_brutto_eink_hh","arbeitsl_geld_2_eink_hh","kindergeld_m_hh","unterhaltsvors_m","jahr","kinderzuschlag_m_vorläufig",,,,"kinderzuschlag_eink_regel","kinderzuschlag_kosten_unterk_m",,"kinderzuschlag_eink_regel","kinderzuschlag_eink_max","kinderzuschlag_eink_spanne","kinderzuschlag_eink_anrechn","kinderzuschlag_temp"
 1,1,1,"TRUE",4,1,3,"FALSE","FALSE",30,0,2000,"FALSE",700,100,"FALSE",0,2,1,1200,644,190,0,2016,140,,,,728,665.28,,1393.28,1563.3,"TRUE",0,140
 1,1,2,"FALSE",4,1,3,"FALSE","FALSE",30,0,2000,"FALSE",700,100,"FALSE",0,2,1,1200,644,190,0,2016,140,,,,728,665.28,,1393.28,1563.3,"TRUE",0,140
 1,1,3,"FALSE",4,1,3,"TRUE","TRUE",2,0,0,"FALSE",700,100,"FALSE",0,2,1,1200,644,190,0,2016,140,,,,728,665.28,,1393.28,1563.3,"TRUE",0,140
