+++ gettsim/tests/test_data/test_dfs_wg.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","kind","kindergeld_anspruch","kaltmiete_m_hh","heizkost_m","alleinerziehend","alter","immobilie_baujahr_hh","mietstufe","bruttolohn_m","ges_rente_m","_ertragsanteil","arbeitsl_geld_m","sonstig_eink_m","unterhaltsvors_m","elterngeld_m","brutto_eink_1","brutto_eink_4","brutto_eink_5","brutto_eink_6","eink_st_m","rentenv_beitr_m","ges_krankenv_beitr_m","behinderungsgrad","jahr","wohngeld_basis_hh",,"_wohngeld_abzüge","_wohngeld_sonstiges_eink","wginc_tu","wg_incdeduct","wgY","a","b","c","max_rent","min_rent","Y","M"
+"hh_id","tu_id","p_id","kind","kindergeld_anspruch","kaltmiete_m_hh","heizkost_m","alleinerziehend","alter","immobilie_baujahr_hh","mietstufe","bruttolohn_m","ges_rente_m","ertragsanteil","arbeitsl_geld_m","sonstig_eink_m","unterhaltsvors_m","elterngeld_m","brutto_eink_1","brutto_eink_4","brutto_eink_5","brutto_eink_6","eink_st_m","rentenv_beitr_m","ges_krankenv_beitr_m","behinderungsgrad","jahr","wohngeld_basis_hh",,"_wohngeld_abzüge","_wohngeld_sonstiges_eink","wginc_tu","wg_incdeduct","wgY","a","b","c","max_rent","min_rent","Y","M"
 1,1,15,"FALSE","FALSE",700,100,"FALSE",45,1995,3,2000,0,0,0,0,0,0,0,25000,1000,0,1000,100,80,0,2016,58.49,,0.3,2166.67,2166.67,100,1446.7,0.02,0.00038,0.000083,563,70,1446.7,563
 1,1,16,"FALSE","FALSE",700,100,"FALSE",46,1995,3,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2016,58.49,,0.3,0,2166.67,100,1446.7,0.02,0.00038,0.000083,563,70,1446.7,563
 1,1,17,"TRUE","TRUE",700,100,"FALSE",9,1995,3,200,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2016,58.49,,0.3,0,2166.67,100,1446.7,0.02,0.00038,0.000083,563,70,1446.7,563
