+++ gettsim/tests/test_wohngeld.py
@@ -24,7 +24,7 @@ INPUT_COLS = [
     "mietstufe",
     "bruttolohn_m",
     "ges_rente_m",
-    "_ertragsanteil",
+    "ertragsanteil",
     "elterngeld_m",
     "arbeitsl_geld_m",
     "sonstig_eink_m",
@@ -62,7 +62,7 @@ def test_wg(input_data, year, column):
         "elterngeld_m",
         "arbeitsl_geld_m",
         "unterhaltsvors_m",
-        "_ertragsanteil",
+        "ertragsanteil",
         "brutto_eink_1",
         "brutto_eink_4",
         "brutto_eink_5",
@@ -102,7 +102,7 @@ def input_data_households():
             "kindergeld_anspruch": False,
             "bruttolohn_m": 0,
             "ges_rente_m": 0,
-            "_ertragsanteil": 0,
+            "ertragsanteil": 0,
             "elterngeld_m": 0,
             "mietstufe": 0,
             "arbeitsl_geld_m": 0,
@@ -131,7 +131,7 @@ def test_increasing_hh_size(input_data_households, year, mietstufe):
         "elterngeld_m",
         "arbeitsl_geld_m",
         "unterhaltsvors_m",
-        "_ertragsanteil",
+        "ertragsanteil",
         "brutto_eink_1",
         "brutto_eink_4",
         "brutto_eink_5",
